database dropped and newly created with implementing all previous changes
