--- db/migrate/20220611131410_create_users.rb
@@ -1,12 +1,0 @@
-class CreateUsers < ActiveRecord::Migration[7.0]
-    def change
-        create_table :users do |t|
-            t.string :name
-            t.string :photo
-            t.text :bio
-            t.integer :posts_counter
-
-            t.timestamps
-        end
-    end
-end
