--- db/migrate/20220611134655_create_likes.rb
@@ -1,8 +1,0 @@
-class CreateLikes < ActiveRecord::Migration[7.0]
-    def change
-        create_table :likes do |t|
-
-            t.timestamps
-        end
-    end
-end
