--- db/migrate/20220611135222_create_comments.rb
@@ -1,9 +1,0 @@
-class CreateComments < ActiveRecord::Migration[7.0]
-    def change
-        create_table :comments do |t|
-            t.text :Text
-
-            t.timestamps
-        end
-    end
-end
