--- db/migrate/20220611140222_add_user_ref_to_comments.rb
@@ -1,5 +1,0 @@
-class AddUserRefToComments < ActiveRecord::Migration[7.0]
-    def change
-        add_reference :comments, :author, null: false, foreign_key: { to_table: :users }
-    end
-end
