--- db/migrate/20220611140607_add_user_ref_to_likes.rb
@@ -1,5 +1,0 @@
-class AddUserRefToLikes < ActiveRecord::Migration[7.0]
-    def change
-        add_reference :likes, :author, null: false, foreign_key: { to_table: :users }
-    end
-end
