--- db/migrate/20220611140800_add_user_ref_to_posts.rb
@@ -1,5 +1,0 @@
-class AddUserRefToPosts < ActiveRecord::Migration[7.0]
-    def change
-        add_reference :posts, :author, null: false, foreign_key: { to_table: :users }
-    end
-end
