--- db/migrate/20220611140930_add_post_ref_to_comments.rb
@@ -1,5 +1,0 @@
-class AddPostRefToComments < ActiveRecord::Migration[7.0]
-    def change
-        add_reference :comments, :post, null: false, foreign_key: true
-    end
-end
