+++ db/migrate/20220615193704_add_user_ref_to_posts.rb
@@ -1,0 +1,5 @@
+class AddUserRefToPosts < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :posts, :user, null: false, foreign_key: { to_table: :users }
+  end
+end
