+++ db/migrate/20220615194758_add_user_ref_to_comments.rb
@@ -1,0 +1,5 @@
+class AddUserRefToComments < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :comments, :user, null: false, foreign_key: { to_table: :users }
+  end
+end
