+++ db/migrate/20220615194808_add_user_ref_to_likes.rb
@@ -1,0 +1,5 @@
+class AddUserRefToLikes < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :likes, :user, null: false, foreign_key: { to_table: :users }
+  end
+end
