+++ db/migrate/20220615200143_add_post_ref_to_comments.rb
@@ -1,0 +1,5 @@
+class AddPostRefToComments < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :comments, :post, null: false, foreign_key: true
+  end
+end
