+++ db/migrate/20220615200333_add_post_ref_to_likes.rb
@@ -1,0 +1,5 @@
+class AddPostRefToLikes < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :likes, :post, null: false, foreign_key: true
+  end
+end
