+++ db/schema.rb
@@ -10,27 +10,27 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_06_11_141130) do
+ActiveRecord::Schema[7.0].define(version: 2022_06_15_200333) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
   create_table "comments", force: :cascade do |t|
-    t.text "Text"
+    t.text "text"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.bigint "author_id", null: false
+    t.bigint "user_id", null: false
     t.bigint "post_id", null: false
-    t.index ["author_id"], name: "index_comments_on_author_id"
     t.index ["post_id"], name: "index_comments_on_post_id"
+    t.index ["user_id"], name: "index_comments_on_user_id"
   end
 
   create_table "likes", force: :cascade do |t|
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.bigint "author_id", null: false
+    t.bigint "user_id", null: false
     t.bigint "post_id", null: false
-    t.index ["author_id"], name: "index_likes_on_author_id"
     t.index ["post_id"], name: "index_likes_on_post_id"
+    t.index ["user_id"], name: "index_likes_on_user_id"
   end
 
   create_table "posts", force: :cascade do |t|
@@ -40,8 +40,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_06_11_141130) do
     t.integer "likes_counter"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.bigint "author_id", null: false
-    t.index ["author_id"], name: "index_posts_on_author_id"
+    t.bigint "user_id", null: false
+    t.index ["user_id"], name: "index_posts_on_user_id"
   end
 
   create_table "users", force: :cascade do |t|
@@ -54,8 +54,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_06_11_141130) do
   end
 
   add_foreign_key "comments", "posts"
-  add_foreign_key "comments", "users", column: "author_id"
+  add_foreign_key "comments", "users"
   add_foreign_key "likes", "posts"
-  add_foreign_key "likes", "users", column: "author_id"
-  add_foreign_key "posts", "users", column: "author_id"
+  add_foreign_key "likes", "users"
+  add_foreign_key "posts", "users"
 end
